requirements.txt
apache-airflow==2.7.0
pandas==2.1.0
psycopg2-binary==2.9.7
pyarrow==13.0.0
SQLAlchemy==1.4.49
apache-airflow-providers-postgres
python-decouple
python-dotenv
